feat(sentry_ratelimit_audit): add hide-dsn and fix
This changes the script to enable a "fix" mode where it sets the
ratelimit for project DSNs that don't have one set.

It also adds a "hide-dsn" mode where it masks the DSNs for putting in
issue comments and other public places. It still lets you distinguish
two DSNs, but the bulk of the DSN id is masked.